requirements.txt
streamlit
tensorflow
mtcnn
opencv-python
numpy
pillow
pandas
